Added internal cache to preferences backend to improve performance.

+++ Kernel/System/CustomerUser/Preferences/DB.pm
@@ -1,8 +1,8 @@
 # --
 # Kernel/System/CustomerUser/Preferences/DB.pm - some customer user functions
-# Copyright (C) 2001-2009 OTRS AG, http://otrs.org/
+# Copyright (C) 2001-2012 OTRS AG, http://otrs.org/
 # --
-# $Id: DB.pm,v 1.20 2009-10-07 20:41:50 martin Exp $
+# $Id: DB.pm,v 1.21 2012-03-14 13:20:35 mh Exp $
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (AGPL). If you
@@ -14,8 +14,10 @@ package Kernel::System::CustomerUser::Preferences::DB;
 use strict;
 use warnings;
 
+use Kernel::System::CacheInternal;
+
 use vars qw(@ISA $VERSION);
-$VERSION = qw($Revision: 1.20 $) [1];
+$VERSION = qw($Revision: 1.21 $) [1];
 
 sub new {
     my ( $Type, %Param ) = @_;
@@ -28,6 +30,12 @@ sub new {
     for (qw(DBObject ConfigObject LogObject)) {
         $Self->{$_} = $Param{$_} || die "Got no $_!";
     }
+
+    $Self->{CacheInternalObject} = Kernel::System::CacheInternal->new(
+        %{$Self},
+        Type => 'CustomerUserPreferencesDB',
+        TTL  => 60 * 60 * 3,
+    );
 
     # preferences table data
     $Self->{PreferencesTable} = $Self->{ConfigObject}->Get('CustomerPreferences')->{Params}->{Table}
@@ -42,21 +50,29 @@ sub new {
         = $Self->{ConfigObject}->Get('CustomerPreferences')->{Params}->{TableUserID}
         || 'user_id';
 
+    # create cache prefix
+    $Self->{CachePrefix} = 'CustomerUserPreferencesDB'
+        . $Self->{PreferencesTable}
+        . $Self->{PreferencesTableKey}
+        . $Self->{PreferencesTableValue}
+        . $Self->{PreferencesTableUserID};
+
     return $Self;
 }
 
 sub SetPreferences {
     my ( $Self, %Param ) = @_;
 
-    my $UserID = $Param{UserID} || return;
-    my $Key    = $Param{Key}    || return;
-    my $Value = defined( $Param{Value} ) ? $Param{Value} : '';
+    return if !$Param{UserID};
+    return if !$Param{Key};
+
+    my $Value = defined $Param{Value} ? $Param{Value} : '';
 
     # delete old data
     return if !$Self->{DBObject}->Do(
         SQL => "DELETE FROM $Self->{PreferencesTable} WHERE "
             . " $Self->{PreferencesTableUserID} = ? AND $Self->{PreferencesTableKey} = ?",
-        Bind => [ \$UserID, \$Key ],
+        Bind => [ \$Param{UserID}, \$Param{Key} ],
     );
 
     # insert new data
@@ -64,7 +80,12 @@ sub SetPreferences {
         SQL => "INSERT INTO $Self->{PreferencesTable} ($Self->{PreferencesTableUserID}, "
             . " $Self->{PreferencesTableKey}, $Self->{PreferencesTableValue}) "
             . " VALUES (?, ?, ?)",
-        Bind => [ \$UserID, \$Key, \$Value ],
+        Bind => [ \$Param{UserID}, \$Param{Key}, \$Value ],
+    );
+
+    # delete cache
+    $Self->{CacheInternalObject}->Delete(
+        Key => $Self->{CachePrefix} . $Param{UserID},
     );
 
     return 1;
@@ -73,28 +94,38 @@ sub SetPreferences {
 sub GetPreferences {
     my ( $Self, %Param ) = @_;
 
-    my $UserID = $Param{UserID} || return;
-    my %Data;
+    return if !$Param{UserID};
+
+    # read cache
+    my $Cache = $Self->{CacheInternalObject}->Get(
+        Key => $Self->{CachePrefix} . $Param{UserID},
+    );
+    return %{$Cache} if $Cache;
 
     # get preferences
-
     return if !$Self->{DBObject}->Prepare(
         SQL => "SELECT $Self->{PreferencesTableKey}, $Self->{PreferencesTableValue} "
             . " FROM $Self->{PreferencesTable} WHERE $Self->{PreferencesTableUserID} = ?",
-        Bind => [ \$UserID ],
+        Bind => [ \$Param{UserID} ],
     );
+
+    my %Data;
     while ( my @Row = $Self->{DBObject}->FetchrowArray() ) {
         $Data{ $Row[0] } = $Row[1];
     }
 
-    # return data
+    # set cache
+    $Self->{CacheInternalObject}->Set(
+        Key   => $Self->{CachePrefix} . $Param{UserID},
+        Value => \%Data,
+    );
+
     return %Data;
 }
 
 sub SearchPreferences {
     my ( $Self, %Param ) = @_;
 
-    my %UserID;
     my $Key   = $Param{Key}   || '';
     my $Value = $Param{Value} || '';
 
@@ -109,11 +140,12 @@ sub SearchPreferences {
         . $Self->{DBObject}->Quote( $Value, 'Like' ) . "')";
 
     return if !$Self->{DBObject}->Prepare( SQL => $SQL );
+
+    my %UserID;
     while ( my @Row = $Self->{DBObject}->FetchrowArray() ) {
         $UserID{ $Row[0] } = $Row[1];
     }
 
-    # return data
     return %UserID;
 }
